Move component pieces around
