+++ jest.config.js
@@ -22,9 +22,11 @@ module.exports = {
 
   // An array of glob patterns indicating a set of files for which coverage information should be collected
   collectCoverageFrom: [
-    'src/js/**.js',
-    'src/js/**.jsx',
-    'src/js/components/**.jsx'
+    'src/js/*.js',
+    'src/js/*.jsx',
+    'src/js/*.jsx',
+    'src/js/components/**/*.jsx',
+    'src/js/utils/*.jsx'
   ],
 
   // The directory where Jest should output its coverage files
