+++ src/js/components/shepherd-content/shepherd-header/index.jsx
