+++ src/js/components/shepherd-content/styles.js
@@ -1,0 +1,10 @@
+export default function contentStyles(variables) {
+  return {
+    content: {
+      background: variables.shepherdTextBackground,
+      fontSize: 'inherit',
+      outline: 'none',
+      padding: 0
+    }
+  };
+}
